+++ code/game/objects/effects/portals.dm
@@ -8,49 +8,52 @@
 	var/failchance = 5
 	var/obj/item/target = null
 	var/creator = null
-	anchored = 1
-	var/precision = 1 // how close to the portal you will teleport. 0 = on the portal, 1 = adjacent
-	var/can_multitool_to_remove = 0
+	anchored = 1.0
 
 /obj/effect/portal/Bumped(mob/M as mob|obj)
-	teleport(M)
+	spawn(0)
+		src.teleport(M)
+		return
+	return
 
-/obj/effect/portal/New(loc, turf/target, creator=null, lifespan=300)
-	portals += src
-	src.loc = loc
-	src.target = target
-	src.creator = creator
-	if(lifespan > 0)
-		spawn(lifespan)
+/obj/effect/portal/Crossed(AM as mob|obj)
+	spawn(0)
+		src.teleport(AM)
+		return
+	return
+
+/obj/effect/portal/attack_hand(mob/user as mob)
+	spawn(0)
+		src.teleport(user)
+		return
+	return
+
+/obj/effect/portal/New(var/start, var/end, var/delete_after = 300)
+	..()
+	playsound(src, 'sound/effects/phasein.ogg', 25, 1)
+	target = end
+	if(delete_after)
+		spawn(delete_after)
 			qdel(src)
 
 /obj/effect/portal/Destroy()
-	portals -= src
-	if(istype(creator, /obj/item/weapon/hand_tele))
-		var/obj/item/weapon/hand_tele/O = creator
-		O.active_portals--
-	else if(istype(creator, /obj/item/weapon/gun/energy/wormhole_projector))
-		var/obj/item/weapon/gun/energy/wormhole_projector/P = creator
-		P.portal_destroyed(src)
-	creator = null
 	target = null
-	return ..()
+	. = ..()
 
 /obj/effect/portal/proc/teleport(atom/movable/M as mob|obj)
 	if(istype(M, /obj/effect)) //sparks don't teleport
 		return
-	if(M.anchored&&istype(M, /obj/mecha))
+	if (M.anchored&&istype(M, /obj/mecha))
 		return
-	if(!( target ))
+	if (icon_state == "portal1")
+		return
+	if (!( target ))
 		qdel(src)
 		return
-	if(istype(M, /atom/movable))
+	if (istype(M, /atom/movable))
 		if(prob(failchance)) //oh dear a problem, put em in deep space
 			src.icon_state = "portal1"
-			do_teleport(M, locate(rand(5, world.maxx - 5), rand(5, world.maxy -5), 3), 0)
+			var/destination_z = GLOB.using_map.get_transit_zlevel(z)
+			do_teleport(M, locate(rand(TRANSITIONEDGE, world.maxx - TRANSITIONEDGE), rand(TRANSITIONEDGE, world.maxy -TRANSITIONEDGE), destination_z), 0)
 		else
-			do_teleport(M, target, precision) ///You will appear adjacent to the beacon
-
-/obj/effect/portal/attackby(obj/item/A, mob/user)
-	if(istype(A, /obj/item/device/multitool) && can_multitool_to_remove)
-		qdel(src)
+			do_teleport(M, target, 1) ///You will appear adjacent to the beacon
