+++ code/game/objects/items/weapons/stunbaton.dm
@@ -1,55 +1,47 @@
+//replaces our stun baton code with /tg/station's code
 /obj/item/weapon/melee/baton
 	name = "stunbaton"
 	desc = "A stun baton for incapacitating people with."
 	icon_state = "stunbaton"
 	item_state = "baton"
 	slot_flags = SLOT_BELT
-	force = 10
+	force = 15
+	sharp = 0
+	edge = 0
 	throwforce = 7
-	w_class = 3
-	origin_tech = "combat=2"
+	w_class = ITEM_SIZE_NORMAL
+	origin_tech = list(TECH_COMBAT = 2)
 	attack_verb = list("beaten")
-	var/stunforce = 7
-	var/status = 0
-	var/obj/item/weapon/stock_parts/cell/high/bcell = null
-	var/hitcost = 1500
+	var/stunforce = 0
+	var/agonyforce = 60
+	var/status = 0		//whether the thing is on or not
+	var/obj/item/weapon/cell/bcell
+	var/hitcost = 10
 
-/obj/item/weapon/melee/baton/suicide_act(mob/user)
-	user.visible_message("<span class='suicide'>[user] is putting the live [name] in \his mouth! It looks like \he's trying to commit suicide.</span>")
-	return (FIRELOSS)
+/obj/item/weapon/melee/baton/loaded
+	bcell = /obj/item/weapon/cell/device/high
 
 /obj/item/weapon/melee/baton/New()
+	if(ispath(bcell))
+		bcell = new bcell(src)
+		update_icon()
 	..()
-	update_icon()
-	return
 
 /obj/item/weapon/melee/baton/Destroy()
-	if(bcell)
+	if(bcell && !ispath(bcell))
 		qdel(bcell)
 		bcell = null
 	return ..()
 
-/obj/item/weapon/melee/baton/CheckParts(list/parts_list)
-	..()
-	bcell = locate(/obj/item/weapon/stock_parts/cell) in contents
-	update_icon()
-
-/obj/item/weapon/melee/baton/loaded/New() //this one starts with a cell pre-installed.
-	..()
-	bcell = new(src)
-	update_icon()
-	return
-
 /obj/item/weapon/melee/baton/proc/deductcharge(var/chrgdeductamt)
 	if(bcell)
-		if(bcell.charge < (hitcost+chrgdeductamt)) // If after the deduction the baton doesn't have enough charge for a stun hit it turns off.
-			status = 0
-			update_icon()
-			playsound(loc, "sparks", 75, 1, -1)
-		if(bcell.use(chrgdeductamt))
+		if(bcell.checked_use(chrgdeductamt))
 			return 1
 		else
+			status = 0
+			update_icon()
 			return 0
+	return null
 
 /obj/item/weapon/melee/baton/update_icon()
 	if(status)
@@ -59,140 +51,168 @@
 	else
 		icon_state = "[initial(name)]"
 
+	if(icon_state == "[initial(name)]_active")
+		set_light(1.5, 2, "#FF6A00")
+	else
+		set_light(0)
+
 /obj/item/weapon/melee/baton/examine(mob/user)
-	..(user)
+	if(!..(user, 1))
+		return 0
+	examine_cell(user)
+	return 1
+
+// Addition made by Techhead0, thanks for fullfilling the todo!
+/obj/item/weapon/melee/baton/proc/examine_cell(mob/user)
 	if(bcell)
 		to_chat(user, "<span class='notice'>The baton is [round(bcell.percent())]% charged.</span>")
 	if(!bcell)
 		to_chat(user, "<span class='warning'>The baton does not have a power source installed.</span>")
 
-/obj/item/weapon/melee/baton/attackby(obj/item/weapon/W, mob/user, params)
-	if(istype(W, /obj/item/weapon/stock_parts/cell))
-		var/obj/item/weapon/stock_parts/cell/C = W
-		if(bcell)
-			to_chat(user, "<span class='notice'>[src] already has a cell.</span>")
-		else
-			if(C.maxcharge < hitcost)
-				to_chat(user, "<span class='notice'>[src] requires a higher capacity cell.</span>")
-				return
-			if(!user.unEquip(W))
-				return
-			W.loc = src
+/obj/item/weapon/melee/baton/attackby(obj/item/weapon/W, mob/user)
+	if(istype(W, /obj/item/weapon/cell/device))
+		if(!bcell && user.unEquip(W))
+			W.forceMove(src)
 			bcell = W
-			to_chat(user, "<span class='notice'>You install a cell in [src].</span>")
+			to_chat(user, "<span class='notice'>You install a cell into the [src].</span>")
 			update_icon()
-
+		else
+			to_chat(user, "<span class='notice'>[src] already has a cell.</span>")
 	else if(istype(W, /obj/item/weapon/screwdriver))
 		if(bcell)
-			bcell.updateicon()
-			bcell.loc = get_turf(src.loc)
+			bcell.update_icon()
+			bcell.dropInto(loc)
 			bcell = null
 			to_chat(user, "<span class='notice'>You remove the cell from the [src].</span>")
 			status = 0
 			update_icon()
-			return
+	else
 		..()
-	return
 
 /obj/item/weapon/melee/baton/attack_self(mob/user)
+	set_status(!status, user)
+	add_fingerprint(user)
+
+/obj/item/weapon/melee/baton/proc/set_status(var/newstatus, mob/user)
 	if(bcell && bcell.charge > hitcost)
-		status = !status
-		to_chat(user, "<span class='notice'>[src] is now [status ? "on" : "off"].</span>")
-		playsound(loc, "sparks", 75, 1, -1)
+		if(status != newstatus)
+			change_status(newstatus)
+			to_chat(user, "<span class='notice'>[src] is now [status ? "on" : "off"].</span>")
+			playsound(loc, "sparks", 75, 1, -1)
 	else
-		status = 0
+		change_status(0)
 		if(!bcell)
 			to_chat(user, "<span class='warning'>[src] does not have a power source!</span>")
 		else
-			to_chat(user, "<span class='warning'>[src] is out of charge.</span>")
-	update_icon()
-	add_fingerprint(user)
+			to_chat(user,  "<span class='warning'>[src] is out of charge.</span>")
 
-/obj/item/weapon/melee/baton/attack(mob/M, mob/living/user)
+// Proc to -actually- change the status, and update the icons as well.
+// Also exists to ease "helpful" admin-abuse in case an bug prevents attack_self
+// to occur would appear. Hopefully it wasn't necessary.
+/obj/item/weapon/melee/baton/proc/change_status(var/s)
+	if (status != s)
+		status = s
+		update_icon()
+
+/obj/item/weapon/melee/baton/attack(mob/M, mob/user)
 	if(status && (CLUMSY in user.mutations) && prob(50))
-		user.visible_message("<span class='danger'>[user] accidentally hits themself with [src]!</span>", \
-							"<span class='userdanger'>You accidentally hit yourself with [src]!</span>")
-		user.Weaken(stunforce*3)
+		to_chat(user, "<span class='danger'>You accidentally hit yourself with the [src]!</span>")
+		user.Weaken(30)
 		deductcharge(hitcost)
 		return
+	return ..()
 
-	if(isrobot(M))
-		..()
-		return
-	if(!isliving(M))
-		return
+/obj/item/weapon/melee/baton/apply_hit_effect(mob/living/target, mob/living/user, var/hit_zone)
+	if(isrobot(target))
+		return ..()
 
-	var/mob/living/L = M
+	var/agony = agonyforce
+	var/stun = stunforce
+	var/obj/item/organ/external/affecting = null
+	if(ishuman(target))
+		var/mob/living/carbon/human/H = target
+		affecting = H.get_organ(hit_zone)
 
-	if(user.a_intent != I_HARM)
-		if(status)
-			user.do_attack_animation(L)
-			baton_stun(L, user)
+	if(user.a_intent == I_HURT)
+		. = ..()
+		if (!.)	//item/attack() does it's own messaging and logs
+			return 0	// item/attack() will return 1 if they hit, 0 if they missed.
+
+		//whacking someone causes a much poorer electrical contact than deliberately prodding them.
+		stun *= 0.5
+		if(status)		//Checks to see if the stunbaton is on.
+			agony *= 0.5	//whacking someone causes a much poorer contact than prodding them.
 		else
-			L.visible_message("<span class='warning'>[user] has prodded [L] with [src]. Luckily it was off.</span>", \
-							"<span class='warning'>[user] has prodded you with [src]. Luckily it was off</span>")
-			return
+			agony = 0	//Shouldn't really stun if it's off, should it?
+		//we can't really extract the actual hit zone from ..(), unfortunately. Just act like they attacked the area they intended to.
+	else if(!status)
+		if(affecting)
+			target.visible_message("<span class='warning'>[target] has been prodded in the [affecting.name] with [src] by [user]. Luckily it was off.</span>")
+		else
+			target.visible_message("<span class='warning'>[target] has been prodded with [src] by [user]. Luckily it was off.</span>")
 	else
-		if(status)
-			baton_stun(L, user)
-		..()
+		if(affecting)
+			target.visible_message("<span class='danger'>[target] has been prodded in the [affecting.name] with [src] by [user]!</span>")
+		else
+			target.visible_message("<span class='danger'>[target] has been prodded with [src] by [user]!</span>")
+		playsound(loc, 'sound/weapons/Egloves.ogg', 50, 1, -1)
 
+	//stun effects
+	if(status)
+		target.stun_effect_act(stun, agony, hit_zone, src)
+		msg_admin_attack("[key_name(user)] stunned [key_name(target)] with the [src].")
 
-/obj/item/weapon/melee/baton/proc/baton_stun(mob/living/L, mob/user)
-	if(ishuman(L))
-		var/mob/living/carbon/human/H = L
-		if(H.check_shields(0, "[user]'s [name]", src, MELEE_ATTACK)) //No message; check_shields() handles that
-			playsound(L, 'sound/weapons/Genhit.ogg', 50, 1)
-			return
-	user.lastattacked = L
-	L.lastattacker = user
-
-	L.Stun(stunforce)
-	L.Weaken(stunforce)
-	L.apply_effect(STUTTER, stunforce)
-
-	L.visible_message("<span class='danger'>[user] has stunned [L] with [src]!</span>", \
-							"<span class='userdanger'>[user] has stunned you with [src]!</span>")
-	playsound(loc, 'sound/weapons/Egloves.ogg', 50, 1, -1)
-
-	if(isrobot(loc))
-		var/mob/living/silicon/robot/R = loc
-		if(R && R.cell)
-			R.cell.use(hitcost)
-	else
 		deductcharge(hitcost)
 
-	if(ishuman(L))
-		var/mob/living/carbon/human/H = L
-		H.forcesay(hit_appends)
+		if(ishuman(target))
+			var/mob/living/carbon/human/H = target
+			H.forcesay(GLOB.hit_appends)
 
-	add_logs(L, user, "stunned", object="stunbaton")
+	return 0
 
 /obj/item/weapon/melee/baton/emp_act(severity)
 	if(bcell)
-		deductcharge(1000 / severity)
-		if(bcell.reliability != 100 && prob(50/severity))
-			bcell.reliability -= 10 / severity
+		bcell.emp_act(severity)	//let's not duplicate code everywhere if we don't have to please.
 	..()
 
-/obj/item/weapon/melee/baton/wash(mob/user, atom/source)
-	if(bcell)
-		if(bcell.charge > 0 && status == 1)
-			flick("baton_active", source)
-			user.Stun(stunforce)
-			user.Weaken(stunforce)
-			user.stuttering = stunforce
-			deductcharge(hitcost)
-			user.visible_message("<span class='warning'>[user] shocks themself while attempting to wash the active [src]!</span>", \
-								"<span class='userdanger'>You unwisely attempt to wash [src] while it's still on.</span>")
-			playsound(src, "sparks", 50, 1)
-			return 1
-	..()
+// Stunbaton module for Security synthetics
+/obj/item/weapon/melee/baton/robot
+	bcell = null
+	hitcost = 100
 
+// Addition made by Techhead0, thanks for fullfilling the todo!
+/obj/item/weapon/melee/baton/robot/examine_cell(mob/user)
+	to_chat(user, "<span class='notice'>The baton is running off an external power supply.</span>")
 
-//secborg stun baton module
-/obj/item/weapon/melee/baton/loaded/robot
-	hitcost = 1000
+// Override proc for the stun baton module, found in PC Security synthetics
+// Refactored to fix #14470 - old proc defination increased the hitcost beyond
+// usability without proper checks.
+// Also hard-coded to be unuseable outside their righteous synthetic owners.
+/obj/item/weapon/melee/baton/robot/attack_self(mob/user)
+	var/mob/living/silicon/robot/R = isrobot(user) ? user : null // null if the user is NOT a robot
+	update_cell(R) // takes both robots and null
+	if (R)
+		return ..()
+	else	// Stop pretending and get out of your cardborg suit, human.
+		to_chat(user, "<span class='warning'>You don't seem to be able interacting with this by yourself..</span>")
+		add_fingerprint(user)
+	return 0
+
+/obj/item/weapon/melee/baton/robot/attackby(obj/item/weapon/W, mob/user)
+	return
+
+/obj/item/weapon/melee/baton/robot/apply_hit_effect(mob/living/target, mob/living/user, var/hit_zone)
+	update_cell(isrobot(user) ? user : null) // update the status before we apply the effects
+	return ..()
+
+// Updates the baton's cell to use user's own cell
+// Otherwise, if null (when the user isn't a robot), render it unuseable
+/obj/item/weapon/melee/baton/robot/proc/update_cell(mob/living/silicon/robot/user)
+	if (!user)
+		bcell = null
+		set_status(0)
+	else if (!bcell || bcell != user.cell)
+		bcell = user.cell // if it is null, nullify it anyway
 
 //Makeshift stun baton. Replacement for stun gloves.
 /obj/item/weapon/melee/baton/cattleprod
@@ -202,6 +222,8 @@
 	item_state = "prod"
 	force = 3
 	throwforce = 5
-	stunforce = 5
-	hitcost = 3750
+	stunforce = 0
+	agonyforce = 60	//same force as a stunbaton, but uses way more charge.
+	hitcost = 25
+	attack_verb = list("poked")
 	slot_flags = null
