+++ code/game/objects/structures/inflatable.dm
@@ -1,46 +1,64 @@
 /obj/item/inflatable
-	name = "inflatable wall"
-	desc = "A folded membrane which rapidly expands into a large cubical shape on activation."
+	name = "inflatable"
+	w_class = ITEM_SIZE_NORMAL
 	icon = 'icons/obj/inflatable.dmi'
-	icon_state = "folded_wall"
-	w_class = 3
+	var/deploy_path = null
 
 /obj/item/inflatable/attack_self(mob/user)
+	if(!deploy_path)
+		return
 	playsound(loc, 'sound/items/zip.ogg', 75, 1)
-	to_chat(user, "\blue You inflate [src].")
-	var/obj/structure/inflatable/R = new /obj/structure/inflatable(user.loc)
+	to_chat(user, "<span class='notice'>You inflate \the [src].</span>")
+	var/obj/structure/inflatable/R = new deploy_path(user.loc)
 	src.transfer_fingerprints_to(R)
 	R.add_fingerprint(user)
 	qdel(src)
 
-/obj/structure/inflatable
+
+/obj/item/inflatable/wall
 	name = "inflatable wall"
+	desc = "A folded membrane which rapidly expands into a large cubical shape on activation."
+	icon_state = "folded_wall"
+	deploy_path = /obj/structure/inflatable/wall
+
+/obj/item/inflatable/door/
+	name = "inflatable door"
+	desc = "A folded membrane which rapidly expands into a simple door on activation."
+	icon_state = "folded_door"
+	deploy_path = /obj/structure/inflatable/door
+
+/obj/structure/inflatable
+	name = "inflatable"
 	desc = "An inflated membrane. Do not puncture."
 	density = 1
 	anchored = 1
 	opacity = 0
-
 	icon = 'icons/obj/inflatable.dmi'
 	icon_state = "wall"
 
+	var/undeploy_path = null
 	var/health = 50.0
 
-/obj/structure/inflatable/initialize(location)
+/obj/structure/inflatable/wall
+	name = "inflatable wall"
+	undeploy_path = /obj/item/inflatable/wall
+
+/obj/structure/inflatable/New(location)
 	..()
-	air_update_turf(1)
+	update_nearby_tiles(need_rebuild=1)
 
 /obj/structure/inflatable/Destroy()
-	air_update_turf(1)
+	update_nearby_tiles()
 	return ..()
 
 /obj/structure/inflatable/CanPass(atom/movable/mover, turf/target, height=0, air_group=0)
 	return 0
 
-/obj/structure/inflatable/CanAtmosPass(turf/T)
-	return !density
-
 /obj/structure/inflatable/bullet_act(var/obj/item/projectile/Proj)
-	health -= Proj.damage
+	var/proj_damage = Proj.get_structure_damage()
+	if(!proj_damage) return
+
+	health -= proj_damage
 	..()
 	if(health <= 0)
 		deflate(1)
@@ -59,49 +77,17 @@
 				deflate(1)
 				return
 
-/obj/structure/inflatable/blob_act()
-	deflate(1)
-
 /obj/structure/inflatable/attack_hand(mob/user as mob)
 	add_fingerprint(user)
 	return
 
-/obj/structure/inflatable/proc/attack_generic(mob/user as mob, damage = 0)	//used by attack_alien, attack_animal, and attack_slime
-	health -= damage
-	if(health <= 0)
-		user.visible_message("<span class='danger'>[user] tears open [src]!</span>")
-		deflate(1)
-	else	//for nicer text~
-		user.visible_message("<span class='danger'>[user] tears at [src]!</span>")
+/obj/structure/inflatable/attackby(obj/item/weapon/W as obj, mob/user as mob)
+	if(!istype(W) || istype(W, /obj/item/weapon/inflatable_dispenser)) return
 
-/obj/structure/inflatable/attack_alien(mob/user as mob)
-	if(islarva(user))
-		return
-	attack_generic(user, 15)
-
-/obj/structure/inflatable/attack_animal(mob/user as mob)
-	if(!isanimal(user))
-		return
-	var/mob/living/simple_animal/M = user
-	if(M.melee_damage_upper <= 0)
-		return
-	attack_generic(M, M.melee_damage_upper)
-
-/obj/structure/inflatable/attack_slime(mob/user as mob)
-	var/mob/living/carbon/slime/S = user
-	if(!S.is_adult)
-		return
-	attack_generic(user, rand(10, 15))
-
-/obj/structure/inflatable/attackby(obj/item/weapon/W as obj, mob/user as mob, params)
-	if(!istype(W))
-		return
-	if(can_puncture(W))
-		visible_message("<span class='danger'>[user] pierces [src] with [W]!</span>")
-		deflate(1)
-	if(W.damtype == BRUTE || W.damtype == BURN)
-		hit(W.force)
+	if((W.damtype == BRUTE || W.damtype == BURN) && W.can_puncture())
 		..()
+		if(hit(W.force))
+			visible_message("<span class='danger'>[user] pierces [src] with [W]!</span>")
 	return
 
 /obj/structure/inflatable/proc/hit(var/damage, var/sound_effect = 1)
@@ -110,13 +96,11 @@
 		playsound(loc, 'sound/effects/Glasshit.ogg', 75, 1)
 	if(health <= 0)
 		deflate(1)
+		return 1
+	return 0
 
-/obj/structure/inflatable/AltClick()
-	if(usr.stat || usr.restrained())
-		return
-	if(!Adjacent(usr))
-		return
-	deflate()
+/obj/structure/inflatable/CtrlClick()
+	hand_deflate()
 
 /obj/structure/inflatable/proc/deflate(var/violent=0)
 	playsound(loc, 'sound/machines/hiss.ogg', 75, 1)
@@ -126,9 +110,11 @@
 		src.transfer_fingerprints_to(R)
 		qdel(src)
 	else
-		visible_message("[src] slowly deflates.")
+		if(!undeploy_path)
+			return
+		visible_message("\The [src] slowly deflates.")
 		spawn(50)
-			var/obj/item/inflatable/R = new /obj/item/inflatable(loc)
+			var/obj/item/inflatable/R = new undeploy_path(src.loc)
 			src.transfer_fingerprints_to(R)
 			qdel(src)
 
@@ -137,24 +123,21 @@
 	set category = "Object"
 	set src in oview(1)
 
-	if(usr.stat || usr.restrained())
+	if(isobserver(usr) || usr.restrained() || !usr.Adjacent(src))
 		return
 
+	verbs -= /obj/structure/inflatable/verb/hand_deflate
 	deflate()
 
-/obj/item/inflatable/door
-	name = "inflatable door"
-	desc = "A folded membrane which rapidly expands into a simple door on activation."
-	icon = 'icons/obj/inflatable.dmi'
-	icon_state = "folded_door"
-
-/obj/item/inflatable/door/attack_self(mob/user)
-	playsound(loc, 'sound/items/zip.ogg', 75, 1)
-	to_chat(user, "\blue You inflate [src].")
-	var/obj/structure/inflatable/door/R = new /obj/structure/inflatable/door(user.loc)
-	src.transfer_fingerprints_to(R)
-	R.add_fingerprint(user)
-	qdel(src)
+/obj/structure/inflatable/attack_generic(var/mob/user, var/damage, var/attack_verb)
+	health -= damage
+	attack_animation(user)
+	if(health <= 0)
+		user.visible_message("<span class='danger'>[user] [attack_verb] open the [src]!</span>")
+		spawn(1) deflate(1)
+	else
+		user.visible_message("<span class='danger'>[user] [attack_verb] at [src]!</span>")
+	return 1
 
 /obj/structure/inflatable/door //Based on mineral door code
 	name = "inflatable door"
@@ -162,8 +145,8 @@
 	anchored = 1
 	opacity = 0
 
-	icon = 'icons/obj/inflatable.dmi'
 	icon_state = "door_closed"
+	undeploy_path = /obj/item/inflatable/door
 
 	var/state = 0 //closed, 1 == open
 	var/isSwitchingStates = 0
@@ -185,16 +168,10 @@
 		return !opacity
 	return !density
 
-/obj/structure/inflatable/door/CanAtmosPass(turf/T)
-	return !density
-
 /obj/structure/inflatable/door/proc/TryToSwitchState(atom/user)
-	if(isSwitchingStates)
-		return
+	if(isSwitchingStates) return
 	if(ismob(user))
 		var/mob/M = user
-		if(world.time - user.last_bumped <= 60)
-			return //NOTE do we really need that?
 		if(M.client)
 			if(iscarbon(M))
 				var/mob/living/carbon/C = M
@@ -210,26 +187,24 @@
 		Close()
 	else
 		Open()
-	air_update_turf(1)
+	update_nearby_tiles()
 
 /obj/structure/inflatable/door/proc/Open()
 	isSwitchingStates = 1
-	//playsound(loc, 'sound/effects/stonedoor_openclose.ogg', 100, 1)
 	flick("door_opening",src)
 	sleep(10)
-	density = 0
-	opacity = 0
+	set_density(0)
+	set_opacity(0)
 	state = 1
 	update_icon()
 	isSwitchingStates = 0
 
 /obj/structure/inflatable/door/proc/Close()
 	isSwitchingStates = 1
-	//playsound(loc, 'sound/effects/stonedoor_openclose.ogg', 100, 1)
 	flick("door_closing",src)
 	sleep(10)
-	density = 1
-	opacity = 0
+	set_density(1)
+	set_opacity(0)
 	state = 0
 	update_icon()
 	isSwitchingStates = 0
@@ -253,7 +228,6 @@
 			var/obj/item/inflatable/door/R = new /obj/item/inflatable/door(loc)
 			src.transfer_fingerprints_to(R)
 			qdel(src)
-	air_update_turf(1)
 
 /obj/item/inflatable/torn
 	name = "torn inflatable wall"
@@ -261,9 +235,9 @@
 	icon = 'icons/obj/inflatable.dmi'
 	icon_state = "folded_wall_torn"
 
-/obj/item/inflatable/torn/attack_self(mob/user)
-	to_chat(user, "<span class='warning'>The inflatable wall is too torn to be inflated!</span>")
-	add_fingerprint(user)
+	attack_self(mob/user)
+		to_chat(user, "<span class='notice'>The inflatable wall is too torn to be inflated!</span>")
+		add_fingerprint(user)
 
 /obj/item/inflatable/door/torn
 	name = "torn inflatable door"
@@ -271,24 +245,16 @@
 	icon = 'icons/obj/inflatable.dmi'
 	icon_state = "folded_door_torn"
 
-/obj/item/inflatable/door/torn/attack_self(mob/user)
-	to_chat(user, "<span class='warning'>The inflatable door is too torn to be inflated!</span>")
-	add_fingerprint(user)
+	attack_self(mob/user)
+		to_chat(user, "<span class='notice'>The inflatable door is too torn to be inflated!</span>")
+		add_fingerprint(user)
 
 /obj/item/weapon/storage/briefcase/inflatable
 	name = "inflatable barrier box"
 	desc = "Contains inflatable walls and doors."
 	icon_state = "inf_box"
 	item_state = "syringe_kit"
-	max_combined_w_class = 21
-	w_class = 3
-
-/obj/item/weapon/storage/briefcase/inflatable/New()
-	..()
-	new /obj/item/inflatable/door(src)
-	new /obj/item/inflatable/door(src)
-	new /obj/item/inflatable/door(src)
-	new /obj/item/inflatable(src)
-	new /obj/item/inflatable(src)
-	new /obj/item/inflatable(src)
-	new /obj/item/inflatable(src)
+	w_class = ITEM_SIZE_LARGE
+	max_storage_space = DEFAULT_LARGEBOX_STORAGE
+	can_hold = list(/obj/item/inflatable)
+	startswith = list(/obj/item/inflatable/door = 2, /obj/item/inflatable/wall = 3)
